requirements.txt
Django
asgiref
pytz
sqlparse
pillow
django-extensions
django-robots
django-debug-toolbar
django-taggit
django-summernote
django-simple-captcha
django-multi-captcha-admin
mysqlclient
django_compressor
django-compressor
django-htmlmin
